LRCI-3892 Remove unnecessary UpstreamBuilds component
